pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.2.1",
    "openai>=1.62.0",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "plotly>=6.0.0",
    "python-jose[cryptography]>=3.3.0",
    "scipy>=1.15.1",
    "sqlalchemy>=2.0.38",
    "streamlit>=1.42.0",
    "trafilatura>=2.0.0",
    "twilio>=9.4.5",
]
